Complete 01-margin-and-padding-1 project for The Odin Project
In this project I adjusted the padding and margins to accomplish the
desired outcome. I did not edit any of the HTML file and I did not
add or remove any styling from the CSS file. I simply updated the values
of the padding and margins to reach the desired outcome.